requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.0
bcrypt==4.1.1
python-dotenv==1.0.0
WTForms==3.1.0
python-dateutil==2.8.2
google-generativeai>=0.3.2
pytest>=8.4.0
pytest-cov>=7.0.0

